Break examples out to example project
