--- http4s/src/main/scala/geotrellis/server/http4s/auth/package.scala
@@ -1,3 +1,0 @@
-package geotrellis.server.http4s.auth
-
-package object auth {}
